# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EleevateOverseas Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students/me:
    get:
      operationId: getMyStudentProfile
      tags: [students]
      summary: Get my student profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    put:
      operationId: updateMyStudentProfile
      tags: [students]
      summary: Update student profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentProfileBody"
      responses:
        "200":
          description: Updated student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /students/me/dashboard-summary:
    get:
      operationId: getStudentDashboardSummary
      tags: [dashboard]
      summary: Student dashboard summary
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboardSummary"
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all student profiles (admin or consultant)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Student profiles list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentListResponse"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student profile by ID (admin or consultant)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student profile by ID (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Consultants ────────────────────────────────────────────────────────────
...
  /students/me/shortlist:
    get:
      operationId: getShortlist
      tags: [students]
      summary: Get student's shortlisted universities
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Shortlisted universities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/University"
  /students/me/shortlist/ids:
    get:
      operationId: getShortlistIds
      tags: [students]
      summary: Get IDs of shortlisted universities
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Shortlisted university IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /students/me/shortlist/ai-recommend:
    post:
      operationId: aiRecommend
      tags: [students]
      summary: Get AI university recommendations based on student profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: AI recommendations
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendations:
                    type: array
                    items:
                      $ref: "#/components/schemas/AiRecommendation"
  /students/me/shortlist/{universityId}:
    post:
      operationId: toggleShortlist
      tags: [students]
      summary: Toggle university bookmark (add if absent, remove if present)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: universityId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Shortlist toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  shortlisted: { type: boolean }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    StudentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/StudentProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    StudentProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        targetCountries:
          type: array
          items:
            type: string
        studyLevel:
          type: string
          enum: [undergraduate, postgraduate, phd, diploma, certificate]
        budget:
          type: number
        gpa:
          type: number
        ieltsScore:
          type: number
        toeflScore:
          type: number
        greScore:
          type: number
        gmatScore:
          type: number
        preferredIntake:
          type: string
        workExperience:
          type: integer
        nationality:
          type: string
        updatedAt:
          type: string
          format: date-time
      required: [id, userId]
    UpdateStudentProfileBody:
      type: object
      properties:
        targetCountries:
          type: array
          items:
            type: string
        studyLevel:
          type: string
        budget:
          type: number
        gpa:
          type: number
        ieltsScore:
          type: number
        toeflScore:
          type: number
        greScore:
          type: number
        gmatScore:
          type: number
        preferredIntake:
          type: string
        workExperience:
          type: integer
        nationality:
          type: string
...
    StudentDashboardSummary:
      type: object
      properties:
        totalApplications:
          type: integer
        shortlistedUniversities:
          type: integer
        documentsUploaded:
          type: integer
        upcomingDeadlines:
          type: integer
        applicationsByStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
            required: [status, count]
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required: [totalApplications, shortlistedUniversities, documentsUploaded, upcomingDeadlines, applicationsByStatus, recentActivity]
...
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [id, type, message, timestamp]
    University:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        country:
          type: string
        city:
          type: string
        ranking:
          type: integer
        logoUrl:
          type: string
        imageUrl:
          type: string
        website:
          type: string
        description:
          type: string
        acceptanceRate:
          type: number
        avgTuitionUsd:
          type: integer
        featured:
          type: boolean
        programCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, country, city, createdAt]
...
    AiRecommendation:
      type: object
      properties:
        universityId: { type: string }
        matchScore: { type: number }
        reasons:
          type: array
          items: { type: string }
        concern: { type: string, nullable: true }
        university:
          $ref: "#/components/schemas/University"
      required: [universityId, matchScore, reasons]
...